+++ app/assets/src/components/common/HostOrganismSearchBox.jsx
